refactor: create a class representing an indexed blog
this achieves a number of goals:
- general oop goals
- it can be used as a context manager
- we resolve the awkward overloading of db (a xapian index of a blog) and db (the SQL db)

we can't always use the BlogIndex as a context manager. this is because documents are loaded lazily, so the db must be open as long as someone might want to read a document. so the BlogIndex provides get_db() in case we have to manage the lifetime manually; see search.py